% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character}
\usage{
get_forecast.character(address)
}
\arguments{
\item{address}{un vecteur de caracteres representant l adresse}
}
\value{
renvoie un tibble avec des donnees de prevision
}
\description{
get_forecast.character
}
