% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric}
\usage{
get_forecast.numeric(xy)
}
\arguments{
\item{xy}{un vecteur numerique contenant la latitude et la longitude}
}
\value{
renvoie un tibble avec des donnees de prevision
}
\description{
get_forecast.numeric
}
