% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{get_forecast}
\alias{get_forecast}
\title{get_forecast}
\usage{
get_forecast(address)
}
\arguments{
\item{address}{Les coordonnees sous forme d un vecteur numerique de taille 2 (latitude, longitude)
ou une adresse sous forme d un caractere.}
}
\value{
Un tibble contenant les previsions meteorologiques.
}
\description{
get_forecast
}
\examples{
# Exemple avec des coordonnees
get_forecast(c(48.8566, 2.3522))

# Exemple avec une adresse
get_forecast("Place de la Concorde, Paris, France")
}
