% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{get_gps_coordinate}
\usage{
get_gps_coordinate(address)
}
\arguments{
\item{address}{adresse un vecteur de caracteres representant l adresse}
}
\value{
renvoie un vecteur numerique avec la latitude et la longitude
}
\description{
get_gps_coordinate
}
