% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{perform_request}
\alias{perform_request}
\title{perform_request}
\usage{
perform_request(lat, lon)
}
\arguments{
\item{lat}{latitude (numeric)}

\item{lon}{longitude (numeric)}
}
\value{
a tibble with 5 row
}
\description{
perform_request
}
\examples{
perform_request(48.85,2.35)
}
