% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{Reponse JSON contenant les donnees meteorologiques.}
}
\value{
Un tibble avec les colonnes date_heure, temperature_celsius,
temperature_ressentie_celsius, precipitation_proba et precipitation.
}
\description{
unnest_response
}
\examples{
resp <- perform_request(48.85,2.35)
unnest_response(resp)
}
