% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{visu_precipitation}
\alias{visu_precipitation}
\title{visu_precipitation}
\usage{
visu_precipitation(address)
}
\arguments{
\item{address}{a character vector representing the address}
}
\value{
a ggplot object for precipitation visualization
}
\description{
visu_precipitation
}
