% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{visu_temp}
\alias{visu_temp}
\title{visu_temp}
\usage{
visu_temp(address)
}
\arguments{
\item{address}{a character vector representing the address}
}
\value{
a ggplot object for temperature visualization
}
\description{
visu_temp
}
