% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{visu_temp_ressentie}
\alias{visu_temp_ressentie}
\title{visu_temp_ressentie}
\usage{
visu_temp_ressentie(address)
}
\arguments{
\item{address}{a character vector representing the address}
}
\value{
a ggplot object for apparent temperature visualization
}
\description{
visu_temp_ressentie
}
